fix(app): move callback listener up the tree so dependencies are not loaded unnecesarrily
