+++ app/pages/_app.tsx
@@ -1,6 +1,8 @@
 // --- React Methods
 import React, { useEffect } from "react";
 
+import { BroadcastChannel } from "broadcast-channel";
+
 // --- Next Methods
 import { AppProps } from "next/app";
 import Head from "next/head";
@@ -19,10 +21,39 @@ import TagManager from "react-gtm-module";
 const FacebookAppId = process.env.NEXT_PUBLIC_PASSPORT_FACEBOOK_APP_ID || "";
 const GTM_ID = process.env.NEXT_PUBLIC_GOOGLE_TAG_MANAGER_ID || "";
 
-function MyApp({ Component, pageProps }: AppProps) {
+function App({ Component, pageProps }: AppProps) {
   useEffect(() => {
     TagManager.initialize({ gtmId: `${GTM_ID}` });
   }, []);
+
+  if (typeof window !== "undefined") {
+    // pull any search params
+    const queryString = new URLSearchParams(window?.location?.search);
+    // Twitter oauth will attach code & state in oauth procedure
+    const queryError = queryString.get("error");
+    const queryCode = queryString.get("code");
+    const queryState = queryString.get("state");
+
+    // We expect for a queryState like" 'twitter-asdfgh', 'google-asdfghjk'
+    const providerPath = queryState?.split("-");
+    const provider = providerPath ? providerPath[0] : undefined;
+
+    // if Twitter oauth then submit message to other windows and close self
+    if ((queryError || queryCode) && queryState && provider) {
+      // shared message channel between windows (on the same domain)
+      const channel = new BroadcastChannel(`${provider}_oauth_channel`);
+
+      // only continue with the process if a code is returned
+      if (queryCode) {
+        channel.postMessage({ target: provider, data: { code: queryCode, state: queryState } });
+      }
+
+      // always close the redirected window
+      window.close();
+
+      return <div></div>;
+    }
+  }
 
   const facebookSdkScript = (
     <script
@@ -74,4 +105,4 @@ function MyApp({ Component, pageProps }: AppProps) {
   );
 }
 
-export default MyApp;
+export default App;
